+++ backend/api/tasks.js
@@ -11,21 +11,42 @@ const accessManager = require("../security/user-and-access-manager");
 
 /**
  * `implantId` must be the one assigned by the implant itself, NOT the database key.
- * Requires user to be logged in.
  */
 router.get("/tasks/:implantId", accessManager.verifyToken, async (req, res) => {
   const implantId = req.paramString("implantId");
+
   log(`GET /tasks/${implantId}`, "Getting tasks for implant...", levels.DEBUG);
+
   let returnStatus = statusCodes.OK;
   let responseJSON = {};
 
   try {
-    const includeSent = req.query.includeSent === "true";
-    const tasks = await tasksService.getTasksForImplant(implantId, includeSent);
-    responseJSON = {
-      tasks: tasks,
-      errors: [],
-    };
+    if (implantId) {
+      const isAuthed = await accessManager.authZCheck(
+        accessManager.operationType.READ,
+        accessManager.targetEntityType.IMPLANT,
+        implantId,
+        accessManager.accessControlType.READ,
+        req.data.userId
+      );
+
+      if (!isAuthed) {
+        returnStatus = statusCodes.FORBIDDEN;
+        responseJSON = {
+          errors: ["You are not allowed to view this implant!"],
+        };
+      } else {
+        const includeSent = req.query.includeSent === "true";
+        const tasks = await tasksService.getTasksForImplant(
+          implantId,
+          includeSent
+        );
+        responseJSON = {
+          tasks: tasks,
+          errors: [],
+        };
+      }
+    }
   } catch (err) {
     log(`GET /tasks/${implantId}`, err, levels.ERROR);
     returnStatus = statusCodes.INTERNAL_SERVER_ERROR;
@@ -35,7 +56,7 @@ router.get("/tasks/:implantId", accessManager.verifyToken, async (req, res) => {
     };
   }
 
-  return res.status(returnStatus).json(responseJSON);
+  res.status(returnStatus).json(responseJSON);
 });
 
 router.get("/task-types", accessManager.verifyToken, async (req, res) => {
@@ -58,26 +79,32 @@ router.get("/task-types", accessManager.verifyToken, async (req, res) => {
     };
   }
 
-  return res.status(returnStatus).json(responseJSON);
+  res.status(returnStatus).json(responseJSON);
 });
 
 /**
  * Expects req.body to contain:
- * - `name` (string)
- * - `params` (array of *unique* strings)
+ * - name {String}
+ * - params {Array of params}
  */
-router.post(
-  "/task-types",
-  accessManager.verifyToken,
-  accessManager.checkAdmin,
-  async (req, res) => {
-    log("POST /task-types", "Creating a task type...", levels.DEBUG);
-    let response = {
-      taskType: null,
-      errors: [],
-    };
-    let status = statusCodes.OK;
-    try {
+router.post("/task-types", accessManager.verifyToken, async (req, res) => {
+  log("POST /task-types", "Creating a task type...", levels.DEBUG);
+  let response = {
+    taskType: null,
+    errors: [],
+  };
+  let status = statusCodes.OK;
+
+  try {
+    const isAuthed = await accessManager.authZCheck(
+      accessManager.operationType.EDIT,
+      null,
+      null,
+      accessManager.accessControlType.ADMIN,
+      req.data.userId
+    );
+
+    if (isAuthed) {
       const validity = validateTaskType(req.body);
       if (validity.isValid) {
         response.taskType = await tasksService.createTaskType(req.body);
@@ -85,42 +112,62 @@ router.post(
         status = statusCodes.BAD_REQUEST;
         response.errors = validity.errors;
       }
-    } catch (err) {
-      log("POST /task-types", err, levels.ERROR);
-      response.errors = ["Internal Server Error"];
-      status = statusCodes.INTERNAL_SERVER_ERROR;
+    } else {
+      response.errors = ["Not authorised"];
+      status = statusCodes.FORBIDDEN;
     }
-    res.status(status).json(response);
+  } catch (err) {
+    log("POST /task-types", err, levels.ERROR);
+
+    response.errors = ["Internal Server Error"];
+    status = statusCodes.INTERNAL_SERVER_ERROR;
   }
-);
+
+  res.status(status).json(response);
+});
 
 router.post("/tasks", accessManager.verifyToken, async (req, res) => {
   log("POST /tasks", `Setting task ${JSON.stringify(req.body)}`, levels.DEBUG);
   let returnStatus = statusCodes.OK;
   let responseJSON = { errors: [] };
 
-  const validationResult = await validateTask(req.body);
-  if (validationResult.isValid) {
-    try {
-      error = await tasksService.setTask(req.body);
-      if (error) {
-        log("POST /tasks", error, levels.WARN);
+  try {
+    const isAuthed = await accessManager.authZCheck(
+      accessManager.operationType.EDIT,
+      accessManager.targetEntityType.IMPLANT,
+      req.bodyString("implantId"),
+      accessManager.accessControlType.EDIT,
+      req.data.userId
+    );
+
+    if (isAuthed) {
+      const validationResult = await validateTask(req.body);
+      if (validationResult.isValid) {
+        error = await tasksService.setTask(req.body); // TODO May be worth recursively sanitising the body down to just strings/ints/bools and objects built from such.
+        if (error) {
+          log("POST /tasks", error, levels.WARN);
+          returnStatus = statusCodes.BAD_REQUEST;
+          responseJSON.errors = [error];
+        }
+      } else {
         returnStatus = statusCodes.BAD_REQUEST;
-        responseJSON.errors = [error];
+        responseJSON.errors = validationResult.errors;
       }
-    } catch (err) {
-      log("POST /tasks", err, levels.ERROR);
-      returnStatus = statusCodes.INTERNAL_SERVER_ERROR;
-      responseJSON = {
-        errors: ["Internal Server Error"],
-      };
+    } else {
+      returnStatus = statusCodes.FORBIDDEN;
+      responseJSON.errors = [
+        "You are not permitted to assign tasks to this implant!",
+      ];
     }
-  } else {
-    returnStatus = statusCodes.BAD_REQUEST;
-    responseJSON.errors = validationResult.errors;
+  } catch (err) {
+    log("POST /tasks", err, levels.ERROR);
+    returnStatus = statusCodes.INTERNAL_SERVER_ERROR;
+    responseJSON = {
+      errors: ["Internal Server Error"],
+    };
   }
 
-  return res.status(returnStatus).json(responseJSON);
+  res.status(returnStatus).json(responseJSON);
 });
 
 router.delete("/tasks/:taskId", accessManager.verifyToken, async (req, res) => {
@@ -133,16 +180,30 @@ router.delete("/tasks/:taskId", accessManager.verifyToken, async (req, res) => {
   let returnStatus = statusCodes.OK;
 
   try {
-    const task = await tasksService.getTaskById(taskId);
+    const task = await tasksService.getTaskById(taskId); // TODO Is there a security risk in exposing the fact that the task does/doesn't exist, to unauthorised people?
+
     if (task) {
-      if (task.sent) {
+      const isAuthed = await accessManager.authZCheck(
+        accessManager.operationType.EDIT,
+        accessManager.targetEntityType.IMPLANT,
+        req.bodyString("implantId"),
+        accessManager.accessControlType.EDIT,
+        req.data.userId
+      );
+
+      if (isAuthed && !task.sent) {
+        await tasksService.deleteTask(taskId);
+      } else if (isAuthed && task.sent) {
         returnStatus = statusCodes.BAD_REQUEST;
         responseJSON.errors.push(
           "Cannot delete a task that has been sent to an implant."
         );
         log(`DELETE /tasks/${taskId}`, "Task already sent", levels.WARN);
       } else {
-        await tasksService.deleteTask(taskId);
+        returnStatus = statusCodes.FORBIDDEN;
+        responseJSON.errors = [
+          "You are not permitted to delete tasks from this implant!",
+        ];
       }
     } else {
       log(
@@ -157,13 +218,12 @@ router.delete("/tasks/:taskId", accessManager.verifyToken, async (req, res) => {
     log(`DELETE /tasks/${taskId}`, err, levels.ERROR);
   }
 
-  return res.status(returnStatus).json(responseJSON);
+  res.status(returnStatus).json(responseJSON);
 });
 
 router.delete(
   "/task-types/:taskTypeId",
   accessManager.verifyToken,
-  accessManager.checkAdmin,
   async (req, res) => {
     const taskTypeId = req.paramString("taskTypeId");
     log(
@@ -178,15 +238,27 @@ router.delete(
     let returnStatus = statusCodes.OK;
 
     try {
-      const taskType = await tasksService.getTaskTypeById(taskTypeId);
-      if (taskType) {
-        await tasksService.deleteTaskType(taskTypeId);
+      const isAuthed = await accessManager.authZCheck(
+        accessManager.operationType.EDIT,
+        null,
+        null,
+        accessManager.accessControlType.ADMIN,
+        req.data.userId
+      );
+      if (isAuthed) {
+        const taskType = await tasksService.getTaskTypeById(taskTypeId);
+        if (taskType) {
+          await tasksService.deleteTaskType(taskTypeId);
+        } else {
+          log(
+            `DELETE /task-types/${taskTypeId}`,
+            `Task type with ID ${taskTypeId} does not exist`,
+            levels.WARN
+          );
+        }
       } else {
-        log(
-          `DELETE /task-types/${taskTypeId}`,
-          `Task type with ID ${taskTypeId} does not exist`,
-          levels.WARN
-        );
+        returnStatus = statusCodes.FORBIDDEN;
+        responseJSON.errors = ["You are not permitted to delete task types!"];
       }
     } catch (err) {
       returnStatus = statusCodes.INTERNAL_SERVER_ERROR;
